Fixed appended animations with relative targets; added mgr.js
Using `append`, relative animations ("+=100px") have not been based
on the final destination of the previous animation, as they should have.

Instead, the target location was chosen relative to the scroll position
at the moment the new animation was appended. This could have been a
random position somewhere in mid-movement.