Switched cores, now using feature testing
When the window is scrolled, either body or documentElement must be
animated, depending on the browser. So far, both were animated (core B).
I have preferred this over feature testing (core A) because the test
adds 5-25ms to the load time - in addition to the 5-25ms consumed by
the jQuery.documentSize dependency. That is up to 50ms just for loading
the plugin. If it can be reduced, it should be.

Core B had its drawbacks, too. Animating both properties caused
animation callbacks to fire twice. But the duplication seemed
manageable with filtering.

Turns out it isn't manageable. When scroll movements are chained
(`append` flag), the queue has to be reordered in some cases. Because
there were two elements with two queues which even behaved differently
(one element did animate, the other didn't), queue reordering derailed
filtering for duplicates. A fix required complex, brittle code which
could have led to race conditions.

So core B and double animation had to go. The decision comes at a price
of 25ms at load time on the slowest of platforms - much less in capable
browsers -, and it is worth it after all.