Tweaked animation duration in tests to improve test reliability
A duration of just 100ms made tests fail in some browsers, some of the time,
even though the code worked correctly. Animations sometimes did not cover the
full distance and stopped at, say, 99px instead of 100px without completing.
This is a jQuery issue: it happens when the browser window is relatively
large, and the duration short.

Increasing the duration to 250ms for the tests seems to have resolved the
issue.