Fixed tests and environment for IE 9, IE10, and partially for IE8
The tests still fail in IE8. It seems that IE8 can't handle the heavily
asynchronous character of the tests. After running a number of tests, IE8
seems to confuse the order in which setup, tests and teardown are supposed
to be run.

Giving up on unit testing in IE8 for now. The component can be checked
manually in the demo instead.